added a static_cast due to possible loss of conversion warning

+++ sourceFiles/Election.cpp
@@ -58,7 +58,7 @@ void Election::printBotVotesFromClub(const string& botName, const string& clubNa
     }
     else
     {
-        clubIndex = distance(clubs.begin(), clubIter);
+        clubIndex = static_cast<int>(distance(clubs.begin(), clubIter));
     }
 
     auto botIter = electoralVotes.find(botName);
